Remove Obsolete Transport Header Files
- Deleted `StdioTransport.h`, `StreamableHTTPBase.h`, and `StreamableHTTPTransport.h` to streamline the codebase and eliminate outdated transport implementations.
- This cleanup is part of ongoing efforts to modernize the transport layer and improve overall maintainability and organization in the codebase.
