+++ Source/Public/Communication/Transport/ITransport.h
@@ -5,16 +5,7 @@
 #include "ErrorBase.h"
 #include "MessageBase.h"
 #include "Sandbox/IMCP.h"
-
-// Poco Net includes
-#include <Poco/Event.h>
-#include <Poco/Exception.h>
-#include <Poco/Net/NetException.h>
-#include <Poco/Net/ServerSocket.h>
-#include <Poco/Runnable.h>
-#include <Poco/StreamCopier.h>
-#include <Poco/Thread.h>
-#include <Poco/URI.h>
+#include "Utilities/Async/MCPTask.h"
 
 MCP_NAMESPACE_BEGIN
 
@@ -52,6 +43,7 @@ struct TransportSendOptions {
     // authentication pipeline.
     optional<AuthInfo> AuthInfo;
 };
+
 // TODO: @HalcyonOmega End Direct Translated Code
 
 // Transport types for easy selection
