+++ Source/Public/Communication/Transport/StreamableHTTP/StreamableHTTPServer.h
